feat: when commiting add a commit message field to the commit modal
