+++ src/store/reducers/appReducer.js
@@ -13,7 +13,6 @@ const initialState = {
   connectionCheck: true,
   updateAvailablePleaseRefesh: false,
   notification: null,
-  commit: false,
   pendingError: false,
   readOnlyMode: true,
   apiKey: null,
@@ -44,9 +43,6 @@ const appReducer = (state = initialState, action) => {
 
     case appActions.DEACTIVATE_PROGRESS_INDICATOR:
       return u({ showProgressOverlay: false }, state);
-
-    case appActions.COMMIT:
-      return u({ commit: action.payload }, state);
 
     case appActions.SET_GLOBAL_ERROR_MESSAGE:
       return u({ errorMessage: action.payload }, state);
